Modularize the introduced scheme components
